+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fix function definition of `vkVoidFunction`
+
 ## [0.19.0+182] - 2021-06-21
 
 ### Added