+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Update Vulkan Headers to version 183
+
 ### Fixed
 
 - Fix function definition of `vkVoidFunction`