+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update Vulkan Headers to version 184
+- Update Vulkan Headers to version 185
 - Use a `SmallVec` instead of a `Vec` when returning collections
 
 ### Fixed