+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Derive `Hash`, `Eq` on applicable structs
 - Update Vulkan Headers to version 185
 - Use a `SmallVec` instead of a `Vec` when returning collections