+++ CHANGELOG.md
@@ -14,8 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Update Vulkan Headers to version 189
 - Derive `Hash`, `Eq` on applicable structs
-- Update Vulkan Headers to version 185
 - Use a `SmallVec` instead of a `Vec` when returning collections
 
 ### Fixed