+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Update Vulkan Headers to version 189
+- Update Vulkan Headers to version 190
 - Derive `Hash`, `Eq` on applicable structs
 - Use a `SmallVec` instead of a `Vec` when returning collections